+++ components/Params/Sliders/RangeSlider.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import Typography from '@material-ui/core/Typography';
-import Slider from '@material-ui/core/Slider';
+import { Slider, Typography } from '@material-ui/core';
 
 const useStyles = makeStyles({
   root: {
